+++ tests/testExam.js
@@ -124,7 +124,7 @@ const partialCreditTests = [
 ];
 
 describe('Exam assessment', function() {
-    this.timeout(5000);
+    this.timeout(20000);
 
     before('set up testing server', helperServer.before);
     after('shut down testing server', helperServer.after);