+++ webhooks/grading/grading.sql
@@ -3,4 +3,21 @@ UPDATE grading_jobs AS gj
 SET
     grading_started_at = $start_time
 WHERE
-    gj.id = $grading_job_id
+    gj.id = $grading_job_id;
+
+-- BLOCK get_job_details
+SELECT
+    (graded_at IS NOT NULL) AS was_graded,
+    s3_bucket,
+    s3_root_key
+FROM
+    grading_jobs
+WHERE
+    id = $grading_job_id;
+
+-- BLOCK update_job_output
+UPDATE grading_jobs
+SET
+    output = $output
+WHERE
+    id = $grading_job_id;